Added more methods, successfully established db connection
Added more methods in the MaterialDAO, also added methods that calculate
subtotal/totalamount of transactionItem or transaction.

+++ src/main/java/org/example/Material.java
@@ -6,10 +6,10 @@ package org.example;
 // you can write this instead:
 
 // the canonical constructor here can be used for fetching material objects from the database
-public record Material(int materialId, String name, String unitOfMeasure, double price, double stockQuantity) {
+public record Material(int materialId, String name, String unitOfMeasure, double price, int stockQuantity) {
     // this overloaded constructor can be used for instantiating new material objects that are yet to be added
     // to the database
-    public Material(String name, String unitOfMeasure, double price, double stockQuantity) {
+    public Material(String name, String unitOfMeasure, double price, int stockQuantity) {
         this(0, name, unitOfMeasure, price, stockQuantity);
     }
 }
